test: test케이스 작성
